fix: Made CORS setup DRY
centralised after request for CORS in config settings
setup cors in conftest as well
