+++ app.py
@@ -1,5 +1,5 @@
 # from middleware.before_request import auth_middleware
-from config.settings import create_app
+from config.settings import cors_setup, create_app
 import os
 
 os.environ.setdefault("FLASK_CONFIG", "config.local")
@@ -8,6 +8,6 @@ config_module = os.getenv("FLASK_CONFIG")
 
 app = create_app(config_module)
 
-# auth_middleware(app)  # Register middleware
+cors_setup(app)
