+++ config/settings.py
@@ -5,12 +5,6 @@ import models
 import router
 from config import configure_app
 
-cors_config = {
-    "origins": ["*"],  # allow CORS requests from this origin
-    "methods": ["GET", "POST", "PUT", "DELETE"],  # allow these methods
-    "allow_headers": ["Content-Type", "Authorization"],  # allow these headers
-}
-
 
 def create_app(config_module = "config.local"):
     app = Flask(__name__)
@@ -25,3 +19,11 @@ def create_app(config_module = "config.local"):
     return app
 
 
+def cors_setup(app):
+    @app.after_request
+    def add_cors_headers(response):
+        print("Adding CORS headers")
+        response.headers["Access-Control-Allow-Origin"] = "*"
+        response.headers["Access-Control-Allow-Methods"] = "GET, POST, OPTIONS, PUT, DELETE"
+        response.headers["Access-Control-Allow-Headers"] = "Content-Type"
+        return response
