+++ router/admin.py
@@ -36,15 +36,10 @@ def category_detail(category_id):
 @admin_router.after_request
 @jwt_required()
 def log_admin_actions_after_request(response):
+    print("Logging admin actions")
     if request.method in ["PUT", "DELETE", "POST"]:
         log_admin_action_repo(current_user, request)
 
     return response
 
 
-@admin_router.after_request
-def add_cors_headers(response):
-    response.headers["Access-Control-Allow-Origin"] = "*"
-    response.headers["Access-Control-Allow-Methods"] = "GET, POST, OPTIONS, PUT, DELETE"
-    response.headers["Access-Control-Allow-Headers"] = "Content-Type"
-    return response
