+++ router/auth.py
@@ -21,9 +21,3 @@ def get_current_user():
     return get_user_view(current_user)
 
 
-@auth_router.after_request
-def add_cors_headers(response):
-    response.headers["Access-Control-Allow-Origin"] = "*"
-    response.headers["Access-Control-Allow-Methods"] = "GET, POST, OPTIONS, PUT, DELETE"
-    response.headers["Access-Control-Allow-Headers"] = "Content-Type"
-    return response
