+++ router/product.py
@@ -44,11 +44,5 @@ def get_category_detail(category_id):
     return get_category_detail_view(category_id)
 
 
-@products_router.after_request
-def add_cors_headers(response):
-    response.headers["Access-Control-Allow-Origin"] = "*"
-    response.headers["Access-Control-Allow-Methods"] = "GET, POST, OPTIONS, PUT, DELETE"
-    response.headers["Access-Control-Allow-Headers"] = "Content-Type"
-    return response
-        
+
