+++ router/user.py
@@ -16,9 +16,3 @@ def update_profile():
     return update_user_profile_view(current_user, request.json)
 
 
-@user_router.after_request
-def add_cors_headers(response):
-    response.headers["Access-Control-Allow-Origin"] = "*"
-    response.headers["Access-Control-Allow-Methods"] = "GET, POST, OPTIONS, PUT, DELETE"
-    response.headers["Access-Control-Allow-Headers"] = "Content-Type"
-    return response
